Fix: Corrected used type (int) -> (size_t)

+++ src/cfgfile.h
@@ -271,7 +271,7 @@ struct ice_config_tag {
     char *yp_url[MAX_YP_DIRECTORIES];
     int yp_url_timeout[MAX_YP_DIRECTORIES];
     int yp_touch_interval[MAX_YP_DIRECTORIES];
-    int num_yp_directories;
+    size_t num_yp_directories;
 };
 
 typedef struct {
